Media: Remove custom filestorage class to store files in db for simplicity.
